extract useGetTopTracks hook
